+++ switch-vision-unifi2mqtt/classic_port_probe.py
@@ -13,7 +13,7 @@ from urllib.parse import quote
 
 import unifi2mqtt as core
 
-VERSION = "4.0.1"
+VERSION = "4.0.2"
 COUNTER_FIELDS = (
     "rx_bytes",
     "tx_bytes",